requirements.txt
opencv-python
opencv-contrib-python
pillow
pandas
openpyxl
